pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "django-redis>=6.0.0",
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "dnspython>=2.8.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sendgrid>=6.12.4",
    "whitenoise>=6.11.0",
    "openpyxl>=3.1.5",
]
